intel-terminal/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
feedparser==6.0.10
apscheduler==3.10.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
aiosqlite==0.19.0
pydantic[email]==2.5.0
pyjwt==2.11.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
